fix : Voucher 저장 작업 오류
* VoucherRepository
voucherRepository.insert()호출 코드 추가

* VoucherMapper
Voucher 인터페이스의 정적 멤버에서
독립적인 파일로 분리
