+++ src/main/java/org/programmers/devcourse/voucher/Console.java
@@ -13,7 +13,7 @@ import org.programmers.devcourse.voucher.engine.blacklist.BlackList;
 import org.programmers.devcourse.voucher.engine.io.Input;
 import org.programmers.devcourse.voucher.engine.io.Output;
 import org.programmers.devcourse.voucher.engine.voucher.Voucher;
-import org.programmers.devcourse.voucher.engine.voucher.Voucher.VoucherMapper;
+import org.programmers.devcourse.voucher.engine.voucher.VoucherMapper;
 import org.springframework.stereotype.Component;
 
 @Component
